requirements.txt
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
dataclasses==0.8
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver
MarkupSafe==1.1.1
matplotlib
numpy
olefile==0.46
opencv-python==4.5.1.48
pandas==1.1.5
Pillow
pyparsing
python-dateutil
pytz==2021.1
PyYAML==5.4.1
scipy==1.5.4
six
torch==1.7.1
torchvision==0.8.2
tornado
tqdm==4.58.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
